Created Route for getting charities
