+++ package.json
@@ -22,6 +22,7 @@
   "license": "MIT",
   "dependencies": {
     "@auth0/auth0-react": "^1.1.0",
+    "axios": "^0.21.0",
     "body-parser": "^1.19.0",
     "bootstrap": "^4.5.3",
     "concurrently": "^4.1.0",
